// src/app/shared/interface/runtime-feature.interface.ts
export interface RuntimeFeatureInterface {
  task: boolean;
  reporting: boolean;
  emailing: boolean;
  promotion: boolean;
  lead: boolean;
  marketingExpense: boolean;
  matchingGlobal: boolean;
  history: boolean;
  mailbox: boolean;
  agenda: boolean;
  mls: boolean;
  fisher: boolean;
  mortgage: boolean;
  cyborg: boolean;
  polygons: boolean;
  sendEmailOnBehalf: boolean;
  contract: boolean;
  sector: boolean;
  customAttribute: boolean;
  matchingMargin: boolean;
  calendarExport: boolean;
  restriction: boolean;
  suggestion: boolean;
}